refactor: extract session events from AgentWorkspace
